Update README.md
Introduction
This course has a 30%-weighted coursework component which will require you to use Tableau to create a story (a collection of dashboards) to discover commercial insights from a provided dataset. The purpose of the project is for you to demonstrate your ability to create effective data visualisations using Tableau to communicate key insights to senior management.